Refactored the watcher code to make it better testable and added unit tests for the Facebook and Twitter watchers.
