+++ packages/js/src/classic-editor/watchers/helpers/constants.js
@@ -1,0 +1,7 @@
+/**
+ * The amount of time to wait until we start updating the DOM with
+ * the value in the store.
+ *
+ * @type {number}
+ */
+export const SYNC_DEBOUNCE_TIME = 500;
